# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a ws notifications app",
        "title": "Swagger Cyberjin notifications API",
        "contact": {
            "name": "Andrey Baybara",
            "email": "baybaraandrey@gmail.com"
        },
        "license": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/_api/": {
            "get": {
                "description": "Application monitor",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "app-monitor"
                ],
                "summary": "Application monitor",
                "responses": {
                    "200": {}
                }
            }
        },
        "/api/v1/notifications/": {
            "post": {
                "description": "notify client",
                "tags": [
                    "rest-notifications"
                ],
                "summary": "notify client"
            }
        },
        "/ws/v1/notifications/": {
            "get": {
                "description": "notifications api jwt token must be sended after websocet connection established",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ws-notifications"
                ],
                "summary": "notifications api",
                "parameters": [
                    {
                        "description": "jwt token",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.JWTAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ws.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ws.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.JWTAuth": {
            "type": "object",
            "properties": {
                "jwt_token": {
                    "type": "string"
                }
            }
        },
        "ws.Message": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}